Use `broadcat_tx_commit` in place of `broadcast_tx_sync`
`broadcast_tx_commit` guarantees that it will not return until the transaction gets into a block.

`broadcast_tx_sync` wait for the transaction to get into mempool to return, but does not wait until it makes into a block.

We still do getTransactionStatus loop for now, cause it likely will be needed for cross-contract calls after Nearmint deprecation.