feat(docs): auto-publish docs to GitHub Pages
Previously, we were copying the output of `yarn doc` from this
repository and commiting it to the `docs` folder in
https://github.com/nearprotocol/docs. Since `nearprotocol/docs` is a
[docusaurus](https://docusaurus.io/) website, we used
`typedoc-plugin-markdown` here in this repository to build documentation
in the format expected by docusaurus.

Last week, realizing that this `yarn doc` output is an auto-generated
artifact that we don't need to commit into either repository, I looked
into automating this whole "build docs, copy to other repository"
process; see https://github.com/nearprotocol/docs/pull/291. While
attempting to finalize this approach with @amgando, we realized it had
multiple issues:

* After about three days of effort, my script still had some major gaps
  that left the user experience lacking:

  * Building a proper sidebar for docs.nearprotocol.com was going to add
    additional complexity to the process
  * Getting "Edit this page" links to link to the proper spot would
    require extra effort

* Even if we spent time fixing those issues, we weren't going to be able
  to make a user experience for docs.nearprotocol.com that beat linking
  to an external docs website. Hosting docs at a consistent URL just
  doesn't give a big enough UX win to justify the maintenance burden
  this approach involves.

This last point ultimately made us scrap the approach. Why not keep
high-level info on docs.nearprotocol.com, and put API Reference docs
elsewhere? This allows each repository to maintain its own API Reference
documentation, using a design aesthetic familiar to the language
community of that repository. This seems like both a UX improvement and
a maintainability improvement for `nearprotocol/docs`.

So, this commit adds the machinery we need to automatically build and
deploy docs on every push to `master` here in `near-api-js`:

* remove `typedoc-plugin-markdown` package, as docs are no longer being
  exported to `nearprotocol/docs` repo and do not need to be built in
  docusaurus format
* remove configuration for `typedoc` command (`--out docs` is default)
* add github workflow for deploying docs on every commit to `master`
* remove `doc` script, as we only expect to run `typedoc` command via
  the automated workflow. Having an extra line in package.json seems
  like it clutters more than it clarifies, at this point.
* in addition to running `typedoc`, add `docs/.nojekyll` file to tell
  GitHub that this is not a Jekyll site, otherwise all paths that start
  with an underscore throw 404 errors