refactor: improve README
Some grammar & stylistic improvements for the README. Now that this will
also be our landing page for API Reference docs (#289), I thought it was
important to clarify that these instructions are for contributing to
this library itself, not using it. We may also want to add some info
about how to use the library soon.