feat: update borsh-js to v1.0.1
We have been working on a new version of borsh-js, this commit
updates all the files that depend on borsh-js. This update is
a breaking change, since the objects schemas (which are exported)
are completely different from v0.7.0 (before) to v1.0.1 (now).

The main benefit of updating borsh-js is that the new version
is agnostic to the class being serialized, meaning that two
"deep-equal" instances of different classes can be serialized
using the same Schema. This means that near-api-js will be
able to (de)serialize objects created across different
library versions (which currently is not possible).